pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.70.0",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
